Split out encounter charting tests into their own file
